Create blank.yml
A game like mines with a preditor for only the admin. And the mines game, 1 coin is 1 GH cedis, and make a fake withdrawals like a real money when the person press cash out it should pend for 3 minutes and 3 minutes it should complete. When the person  cash out, the person should provide his phone number and his or her network like telecel and MTN for withdrawals. And make a working signals for me the admin